requirements.txt
torch
torchvision
tqdm
numpy
pandas
matplotlib
Pillow
scikit-learn
jupyter
